+++ crates/nostr-relay-pool/src/pool/options.rs
@@ -9,6 +9,7 @@ use crate::relay::RelayFilteringMode;
 /// Relay Pool Options
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct RelayPoolOptions {
+    pub(super) max_relays: Option<usize>,
     pub(super) notification_channel_size: usize,
     pub(super) filtering_mode: RelayFilteringMode,
 }
@@ -16,6 +17,7 @@ pub struct RelayPoolOptions {
 impl Default for RelayPoolOptions {
     fn default() -> Self {
         Self {
+            max_relays: None,
             notification_channel_size: 4096,
             filtering_mode: RelayFilteringMode::default(),
         }
@@ -27,6 +29,13 @@ impl RelayPoolOptions {
     #[inline]
     pub fn new() -> Self {
         Self::default()
+    }
+
+    /// Max relays (default: None)
+    #[inline]
+    pub fn max_relays(mut self, num: Option<usize>) -> Self {
+        self.max_relays = num;
+        self
     }
 
     /// Notification channel size (default: 4096)